tests/wpt/web-platform-tests/offscreen-canvas/pixel-manipulation/2d.imageData.get.order.alpha.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.get.order.alpha</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.imageData.get.order.alpha</h1>
<p class="desc">getImageData() returns A in the fourth component</p>


<script>
var t = async_test("getImageData() returns A in the fourth component");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.fillStyle = 'rgba(0, 0, 0, 0.5)';
ctx.fillRect(0, 0, 100, 50);
var imgdata = ctx.getImageData(0, 0, 10, 10);
_assert(imgdata.data[3] < 200, "imgdata.data[\""+(3)+"\"] < 200");
_assert(imgdata.data[3] > 100, "imgdata.data[\""+(3)+"\"] > 100");

t.done();

});
</script>
